feat: Add extra validations. When MasterDevice is not empty and when pheripherals are runing.

+++ server/src/master_device/controller/peripheral/peripheral.controller.ts
@@ -34,16 +34,32 @@ export class PeripheralController {
   @Delete('/delete/:id')
   async deletePeripheral(@Param('id') id: string, @Res() res) {
     try {
-      const deletedPeripheral = await this.peripheralService.deletePeripheral(
-        id,
-      );
-      res.status(HttpStatus.OK).json({
-        message: 'Peripheral successfully deleted',
-        deletedPeripheral,
-      });
+      const getPeripheral = await this.peripheralService.getPeripheral(id);
+
+      if (!getPeripheral) {
+        res.status(HttpStatus.NOT_FOUND).json({
+          message: 'Peripheral not found.',
+        });
+      } else if (getPeripheral.isDeleted) {
+        res.status(HttpStatus.BAD_REQUEST).json({
+          message: 'Peripheral is already deleted.',
+        });
+      } else if (getPeripheral.status) {
+        res.status(HttpStatus.BAD_REQUEST).json({
+          message: 'Peripheral is online. Turn off the peripheral first.',
+        });
+      } else {
+        const deletedPeripheral = await this.peripheralService.deletePeripheral(
+          id,
+        );
+        res.status(HttpStatus.OK).json({
+          message: 'Peripheral successfully deleted.',
+          deletedPeripheral,
+        });
+      }
     } catch {
       res.status(HttpStatus.INTERNAL_SERVER_ERROR).json({
-        message: 'Error deleting peripheral',
+        message: 'Error deleting peripheral.',
       });
     }
   }
